move demos to their own folder
